Improve handling of root certificates on devices prior to Android 14
This uses the complete set of root certificates from mozillla via the
curl project on pre-14 devices, additionally we now use Conscrypt to
provide TLS 1.2 and 1.3 support on pre-Android 10 devices, which, both
together, improves the behaviour to the level of current Android
versions.

Resolves https://github.com/MarcusWolschon/osmeditor4android/issues/3149
