--- src/main/res/raw/entrust_dv_tls_issuing_rsa_ca_2.crt
@@ -1,36 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIGNTCCBB2gAwIBAgIQCtX1c5Gcd8kyz4XgxpLVPDANBgkqhkiG9w0BAQwFADBf
-MQswCQYDVQQGEwJHQjEYMBYGA1UEChMPU2VjdGlnbyBMaW1pdGVkMTYwNAYDVQQD
-Ey1TZWN0aWdvIFB1YmxpYyBTZXJ2ZXIgQXV0aGVudGljYXRpb24gUm9vdCBSNDYw
-HhcNMjQxMjExMDAwMDAwWhcNMjcxMjEwMjM1OTU5WjBRMQswCQYDVQQGEwJDQTEY
-MBYGA1UEChMPRW50cnVzdCBMaW1pdGVkMSgwJgYDVQQDEx9FbnRydXN0IERWIFRM
-UyBJc3N1aW5nIFJTQSBDQSAyMIIBojANBgkqhkiG9w0BAQEFAAOCAY8AMIIBigKC
-AYEAnq02xRySq9MV6qSoqYIMRIs263ND1bdHIagL3budUXdXu6lHyqB1100S1fBf
-7SARFr70utHKaO9bPPktbRfxKw0NAebc/L3qOa6cjQfYYyhv8Y3a0ws/Cbd8a/xc
-uiCPn4u7yswRKJiUfeoOXO9G7GGzIOv2wGiGCKHEoDnwZa8ZwSRrsJVTt3LOpolR
-S9Sod0etzW46Ly4lHQCgxVBeeSS6hBWIARwflA3SuRentB7W2nxud5G2eupDbsnb
-J6lqtzz2XqJeArjE8/xkKHk9GIhXPhEj3K0kekh4fB0me4c2csMdraSioZ7n1C43
-IfOIlRLmswjRFAvpPImWksZLiCIn/hi+UHX+WG5mtWMIGuqQuOjMQWFnSne3giVT
-b6T28puOUHQ8Lj12sotvNopFpVseN9qOK1u+5CLyDZvzSxUXR5wlRmACzqtYzArn
-9dVwlxkmTu+/ovPsz22al3zY2s66m1lZKuIKqM+Aqd6B4RLs9EksHd+mTv9EUnkv
-VIVlAgMBAAGjggF5MIIBdTAfBgNVHSMEGDAWgBRWc1hklfmSGrASKgRieaFAFYgh
-STAdBgNVHQ4EFgQUjUJJN0C5R5WAmL6puTpr8M2WoYMwDgYDVR0PAQH/BAQDAgGG
-MBIGA1UdEwEB/wQIMAYBAf8CAQAwHQYDVR0lBBYwFAYIKwYBBQUHAwEGCCsGAQUF
-BwMCMBMGA1UdIAQMMAowCAYGZ4EMAQIBMFQGA1UdHwRNMEswSaBHoEWGQ2h0dHA6
-Ly9jcmwuc2VjdGlnby5jb20vU2VjdGlnb1B1YmxpY1NlcnZlckF1dGhlbnRpY2F0
-aW9uUm9vdFI0Ni5jcmwwgYQGCCsGAQUFBwEBBHgwdjBPBggrBgEFBQcwAoZDaHR0
-cDovL2NydC5zZWN0aWdvLmNvbS9TZWN0aWdvUHVibGljU2VydmVyQXV0aGVudGlj
-YXRpb25Sb290UjQ2LnA3YzAjBggrBgEFBQcwAYYXaHR0cDovL29jc3Auc2VjdGln
-by5jb20wDQYJKoZIhvcNAQEMBQADggIBAAR7QD/G11LeMu8WoScjnrBMJeepJaK9
-BjFuCxwF+aUUuKxxptz0xjXjB4J5mZaBGTWcfwpDokObsz7D6pScyaf0/kkzCL/g
-lKoEImdv2T0pUaDk3P1RhH3v/nsbWde9IzssrxikpyQqNG1gAb2GCpJswHdwnNs1
-utKeKTYItq4O1oqXPXglNSyBkaYph8ZzXiQgTtQAaiDA7hGOryUzRzdi6KTZAnBl
-AfQw3GzIw9ZdSEuXLoYjPZ9mhgHhXflkxmInYo9PQpO0gDyx7rfhUkEppoS9sgK0
-oc9qLKSPUxt3Nt9a5icghWCL/ViFAhdU4YPGZB3KgaZiu8zbH2kC8JrdI7rskROX
-b30xUGhtHrCVO8lJzS7df4pGcROi5VYloF6Eyq6MPNVyq5/XumC+Gj+TQkIywrW9
-V0IhH9lO32w1ZENtowwFblXfXDhsR7lB3/vLPUTqcqm6vorvixNZjWJuYP5D2nwb
-q552qYAIvb3K9Yo50/J7mFsLNjzaaqKKjoE7zX0nT6LT5LxqdCYP0pQFRu3vf7DL
-pqzLnF1+ZXlkJ/fgQDgv37DotO8GFZ49DnKX4B5utKrdcheSY3IntjCTtcb96Z3C
-zg9mAR0sBwX86RhZKgZ1s33iqEmJXiP51On/M9UzdWdFnth8f4gIKtKPgO3MxupR
-+CgUsrUKgzo4
------END CERTIFICATE-----
