+++ src/main/res/xml/network_security_config.xml
@@ -3,6 +3,7 @@
     <base-config cleartextTrafficPermitted="true">
         <trust-anchors>
             <certificates src="@raw/isrg_root_x1" />
+            <certificates src="@raw/cacert" />
             <certificates src="system" />
             <certificates src="user" />
         </trust-anchors>
